requirements.txt
pydub==0.25.1
PyYAML==6.0
tqdm==4.62.3
